Working checkpoint

+++ cource_work/Models/Entity/DayCashAmount.cs
@@ -11,7 +11,7 @@ namespace cource_work.Models.Entity
         }
 
         public int DcaId { get; set; }
-        public double TotalDayAmount { get; set; }
+        public Decimal TotalDayAmount { get; set; }
         public DateTime WorkDate { get; set; }
         public int AccId { get; set; }
         public int CrId { get; set; }
